+++ end_to_end_test/test/records_serializer_test.dart
@@ -4,12 +4,13 @@
 
 import 'dart:convert';
 
+import 'package:end_to_end_test/errors_matchers.dart';
 import 'package:end_to_end_test/records.dart';
 import 'package:end_to_end_test/serializers.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group(SerializableRecordValue, () {
+  group('$SerializableRecordValue with no record value', () {
     var data = SerializableRecordValue((b) => b..value = 1);
     var serialized = json.decode(json.encode([
       'SerializableRecordValue',
@@ -23,6 +24,43 @@ void main() {
 
     test('can be deserialized', () {
       expect(serializers.deserialize(serialized), data);
+    });
+  });
+
+  group('$SerializableRecordValue with a record value', () {
+    var data = SerializableRecordValue((b) => b
+      ..value = 1
+      ..record = (1, 2));
+    var serialized = json.decode(json.encode([
+      'SerializableRecordValue',
+      'value',
+      1,
+      'record',
+      [1, 2],
+    ])) as Object;
+    var serializersWithCustomSerializer =
+        (serializers.toBuilder()..add(RecordOfIntIntSerializer())).build();
+
+    test('gives advice about custom serializer on failure to serialize', () {
+      expect(
+          () => serializers.serialize(data),
+          throwsA(isErrorContaining(
+              'record types are not automatically serializable')));
+    });
+
+    test('gives advice about custom serializer on failure to deserialize', () {
+      expect(
+          () => serializers.deserialize(serialized),
+          throwsA(isErrorContaining(
+              'record types are not automatically serializable')));
+    });
+
+    test('can be serialized with custom serializer', () {
+      expect(serializersWithCustomSerializer.serialize(data), serialized);
+    });
+
+    test('can be deserialized with custom deserializer', () {
+      expect(serializersWithCustomSerializer.deserialize(serialized), data);
     });
   });
 }